// web/src/api/models/TypesConntrackdStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesConntrackdStatus
 */
export interface TypesConntrackdStatus {
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof TypesConntrackdStatus
     */
    by_proto?: { [key: string]: number; };
    /**
     * 
     * @type {number}
     * @memberof TypesConntrackdStatus
     */
    entries: number;
    /**
     * 
     * @type {boolean}
     * @memberof TypesConntrackdStatus
     */
    running: boolean;
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof TypesConntrackdStatus
     */
    tcp_states?: { [key: string]: number; };
}

/**
 * Check if a given object implements the TypesConntrackdStatus interface.
 */
export function instanceOfTypesConntrackdStatus(value: object): value is TypesConntrackdStatus {
    if (!('entries' in value) || value['entries'] === undefined) return false;
    if (!('running' in value) || value['running'] === undefined) return false;
    return true;
}

export function TypesConntrackdStatusFromJSON(json: any): TypesConntrackdStatus {
    return TypesConntrackdStatusFromJSONTyped(json, false);
}

export function TypesConntrackdStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesConntrackdStatus {
    if (json == null) {
        return json;
    }
    return {
        
        'by_proto': json['by_proto'] == null ? undefined : json['by_proto'],
        'entries': json['entries'],
        'running': json['running'],
        'tcp_states': json['tcp_states'] == null ? undefined : json['tcp_states'],
    };
}

export function TypesConntrackdStatusToJSON(json: any): TypesConntrackdStatus {
    return TypesConntrackdStatusToJSONTyped(json, false);
}

export function TypesConntrackdStatusToJSONTyped(value?: TypesConntrackdStatus | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'by_proto': value['by_proto'],
        'entries': value['entries'],
        'running': value['running'],
        'tcp_states': value['tcp_states'],
    };
}